Only create issues for the current file
verilator seems to be able to report issues on other files in the hierarchy when parsing one file.

I’m not sure if Nova supports reporting issue for another file than the currentone but I made a request for helkp in the Nova forums (https://devforum.nova.app/t/creating-issues-for-another-file-than-the-one-being-parsed/2044).

Until then, we should at least make sure we only report issues for the current file.